+++ packages/adminui-vue/ub-components.js
@@ -2,8 +2,8 @@ module.exports = {
   install (Vue) {
     const components = [
       require('./components/controls/UbSelectEntity.vue').default,
-      require('./components/controls/UbFormRow.vue').default,
-      require('./components/controls/UbErrorWrap.vue').default,
+      require('./components/controls/UFormRow.vue').default,
+      require('./components/controls/UErrorWrap.vue').default,
       require('./components/controls/UbSelectEnum.vue').default
     ]
     for (const component of components) {